// src/enums/enums.ts
/* eslint no-shadow: ["error", { "allow": ["SortingPreference", "PhoneColor", "Manufacturer"] }] */
export enum SortingPreference {
  MOST_SOLD = 'MOST_SOLD',
  DATE = 'DATE',
  PROMOTION = 'PROMOTION',
}

export enum PhoneColor {
  BLACK = 'ZWART',
  WHITE = 'WIT',
  GREY = 'GRIJS',
}

export enum Manufacturer {
  SAMSUNG = 'Samsung',
  OPPO = 'OPPO',
  APPLE = 'Apple',
  MOTOROLA = 'Motorola'
}